Exploratory Data Analysis on Student Performance
For this project, we will use a dataset of test scores from 1000 students found on data.world.com. Students' performances on the tests vary considering the other factors included. We will then used the information from the dataset to determine if there is any correlation between the exam scores and the other factors in the dataset. We will use Matplotlib, Jupyter Notebook, and other resources to clean the data and test the different datasets in order to find how the factors affect the student’s exam scores. After the analysis was complete, we will be able to answer questions we have about the data and be able to support it with statistical data.

Team Members: Fana Wachamo, Charne Jewell-Hart, Debora Rogers, and Amon Thomas